pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yolo_inspection_kit"
version = "1.0.0"
description = "A lightweight YOLO-based inspection toolkit for product quality checking and packaging verification"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "MorganChen", email = "cjfmail@foxmail.com"},
]
keywords = ["yolo", "detection", "inspection", "quality-check", "computer-vision"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
]

dependencies = [
    "ultralytics>=8.0.0",
    "opencv-python>=4.5.0",
    "numpy>=1.20.0",
    "pillow>=8.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=22.0",
    "flake8>=4.0",
    "mypy>=0.990",
]
all = [
    "pytest>=7.0",
    "black>=22.0",
    "flake8>=4.0",
    "mypy>=0.990",
]

[project.urls]
Homepage = "https://github.com/Json031/yolo_inspection_kit"
Documentation = "https://github.com/Json031/yolo_inspection_kit"
Repository = "https://github.com/Json031/yolo_inspection_kit.git"
Issues = "https://github.com/Json031/yolo_inspection_kit/issues"

[tool.setuptools]
packages = ["yolo_inspection_kit"]

[tool.setuptools.package-data]
yolo_inspection_kit = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
